backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
faker
sqlalchemy
requests